d-ai-dynamic-ui/pyproject.toml
[project]
name = "d-ai-dynamic-ui"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-identity>=1.23.0",
    "jinja2>=3.1.6",
    "openai>=1.78.1",
    "python-dotenv>=1.1.0",
    "python-fasthtml>=0.12.16",
]
